+++ Changes.md
@@ -6,6 +6,7 @@ Version 403.1.1 - TBR
 1. Fix 'Import a format_grid course into a format_topics course fails and produces "undefined method"' - #193,
    thanks to @Syxton(Matt Davidson).
 2. Fix 'Sections showing beyond the number set in the settings', related to #196.
+3. Improved backup and restore logic.
 
 Version 403.1.0 - 15/10/2023
 ----------------------------